requirements.txt
alembic==1.14.0
colorama==0.4.6
Faker==20.1.0
greenlet==3.1.1
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
prompt_toolkit==3.0.48
psycopg2-binary==2.9.9
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
rich==13.9.4
six==1.16.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
wcwidth==0.2.13
